requirements.txt
Django==1.11.5
django-cors-headers==2.1.0
django-leaflet==0.22.0
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
psycopg2==2.7.3.1
pytz==2017.2
